refactor: 클래스 재배치 및 패키지 재구성
