+++ src/main/java/racingcar/Application.java
@@ -1,10 +1,10 @@
 package racingcar;
 
-import racingcar.racinggame.Game;
+import racingcar.controller.GameController;
 
 public class Application {
 	public static void main(String[] args) {
-		Game game = new Game();
+		GameController game = new GameController();
 		game.start();
 	}
 }
