+++ src/main/java/racingcar/view/InputView.java
@@ -1,16 +1,16 @@
-package racingcar.user;
+package racingcar.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import racingcar.racinggame.Validation;
+import racingcar.model.Validation;
 
-public class Input {
+public class InputView {
 	private static final String CAR_NAME_INPUT_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
 	private static final String ROUND_INPUT_MESSAGE = "시도할 횟수는 몇회인가요?";
 	private static final String COMMA = ",";
 
 	private final Validation validation;
 
-	public Input() {
+	public InputView() {
 		validation = new Validation();
 	}
